+++ svelte.config.js
@@ -1,9 +1,8 @@
 import adapter from '@sveltejs/adapter-node';
 import { vitePreprocess } from '@sveltejs/vite-plugin-svelte';
-
 const config = {
 	preprocess: vitePreprocess(),
-	kit: { adapter: adapter() },
+	kit: { adapter: adapter(), experimental: { remoteFunctions: true } },
 	optimizeDeps: {
 		exclude: ['output']
 	}
